Initial commit: rework navbar
